requirements.txt

flasgger==0.9.5
Flask==2.0.3
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
python-dotenv==0.19.2
requests==2.27.1
pytz==2022.1
PyMySQL==1.0.2
